server/learnapp/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-extensions==2.2.9
django-simple-history==2.10.0
psycopg2==2.8.5
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
